pyproject.toml
[tool.poetry]
name = "fast-server"
version = "0.1.0"
description = ""
authors = ["hoanghuumanh54 <hoanghuumanh54@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sentence-transformers = "^2.2.2"
pinecone-client = "^2.2.1"
openai = "^0.27.6"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
pydantic = {extras = ["email"], version = "^1.10.7"}
fastapi = "0.65.2"
websockets = "^11.0.3"
sqlalchemy = "1.3.16"
python-jose = "^3.3.0"
passlib = "^1.7.4"
psycopg2-binary = "^2.9.6"
alembic = "^1.10.4"
bcrypt = "^4.0.1"
pyhumps = "3.7.1"
google = "^3.0.0"
google-api-python-client = "^2.86.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
boto3 = "^1.26.134"
gunicorn = "^20.1.0"
python-docx = "^0.8.11"
rq = "^1.14.1"
langdetect = "^1.0.9"
pytesseract = "^0.3.10"
more-itertools = "^9.1.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"
ipykernel = "^6.23.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
